aksd: extract privacyTelemetryInitializer to side-effect-free module
Move the function from analyticsSetup.ts (which has module-load side
effects when REACT_APP_APPINSIGHTS_CONNECTION_STRING is set) to a new
lib/analyticsPrivacy.ts. The unit test imports from there so it never
risks pulling AppInsights initialization into the test's module graph.

Addresses Copilot review feedback on PR #626.
